day11/requirements.txt
streamlit>=1.35.0
huggingface_hub>=0.24.0
python-dotenv>=1.0.0
